# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Servidores",
        "description": "API Servidores",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/unidades": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Lista todas as unidades com paginação",
                "description": "Retorna uma lista paginada de todas as unidades.",
                "operationId": "0c9baef10d156dabf3eeb1740c369edf",
                "responses": {
                    "200": {
                        "description": "Lista de unidades retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UnidadeResource"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Nenhuma unidade encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Cria uma nova unidade",
                "description": "Cria um novo registro de unidade.",
                "operationId": "630809295267e4bd587485d8a29519a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUnidadeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unidade criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados de entrada inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/unidades/{id}": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Busca uma unidade por ID",
                "description": "Retorna os detalhes de uma unidade específica com base no seu ID.",
                "operationId": "460b6c13458ccbe2b8584cd7097d9a0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade a ser buscada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Atualiza uma unidade existente",
                "description": "Atualiza os dados de uma unidade específica com base no seu ID.",
                "operationId": "323706e2f822d031fdeebebd9013f931",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnidadeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados de entrada inválidos"
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Exclui uma unidade por ID",
                "description": "Remove uma unidade específica com base no seu ID.",
                "operationId": "a52327e7319d99b373e840cef59c1afa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Unidade excluída com sucesso"
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/unidades/{id}/enderecos": {
            "post": {
                "tags": [
                    "Unidades",
                    "Endereços"
                ],
                "summary": "Adiciona um endereço a uma unidade existente",
                "description": "Adiciona um novo endereço à unidade especificada.",
                "operationId": "3dcd2f478c7f1793efd9e2ce5eddd154",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade para adicionar o endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddEnderecoToUnidadeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço adicionado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AddEnderecoToUnidadeRequest": {
                "required": [
                    "end_tipo_logradouro",
                    "end_logradouro",
                    "end_numero",
                    "end_bairro",
                    "cid_id"
                ],
                "properties": {
                    "end_tipo_logradouro": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "end_logradouro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "end_numero": {
                        "type": "integer"
                    },
                    "end_bairro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "cid_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "StoreUnidadeRequest": {
                "required": [
                    "unid_nome",
                    "unid_sigla"
                ],
                "properties": {
                    "unid_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "unid_sigla": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "endereco_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateUnidadeRequest": {
                "required": [],
                "properties": {
                    "unid_nome": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "unid_sigla": {
                        "type": "string",
                        "maxLength": 50,
                        "nullable": true
                    },
                    "endereco_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UnidadeResource": {
                "properties": {
                    "unid_id": {
                        "type": "integer"
                    },
                    "unid_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "unid_sigla": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "enderecos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "end_id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}